+++ NEWS.md
@@ -6,7 +6,8 @@
 
 ## Bug fixes
 
-* Fixed the bug of `summarise.tbl_ts()` and `tsummarise.tbl_ts()`, when calling functions with no parameters (like `dplyr::n()`. (@mitchelloharawild)
+* Fixed the bug of `summarise.tbl_ts()` and `tsummarise.tbl_ts()`, when calling functions with no parameters like `dplyr::n()`. (@mitchelloharawild)
+* One grouping level should be dropped for the consistency with `dplyr::summarise()` for a grouped data frame. (@mitchelloharawild)
 
 # tsibble 0.1.0